Update Debt Calc 1.0.py
Had to change the months_paying to int, so it would show as a whole number.

+++ Debt Calc 1.0.py
@@ -40,7 +40,7 @@ else:
     else:
         income = int(input("How much are you putting towards your debt every month?"))
 
-        months_paying = total_debt / income
+        months_paying = int(total_debt / income)
 
 # Grabbing current month/year to be used later
 current_year = datetime.date.today().year
